refactor: move dataform to crud viewers
